+++ loansum-fpml/src/main/java/com/syndloanhub/loansum/fpml/RepaymentExporter.java
@@ -1,7 +1,5 @@
 package com.syndloanhub.loansum.fpml;
 
-import java.math.BigDecimal;
-
 import com.syndloanhub.loansum.fpml.v5_11.confirmation.LoanContract;
 import com.syndloanhub.loansum.fpml.v5_11.confirmation.LoanContractReference;
 import com.syndloanhub.loansum.fpml.v5_11.confirmation.Repayment;
@@ -11,7 +9,7 @@ public class RepaymentExporter {
   static public Repayment convert(com.syndloanhub.loansum.product.facility.Repayment event, LoanContract contract) {
     Repayment fpml = FpMLHelper.factory.createRepayment();
     fpml.setAmount(FpMLHelper.convert(event.getAmount()));
-    fpml.setPrice(BigDecimal.valueOf(event.getPrice()));
+    fpml.setPrice(event.getPrice());
     fpml.setEffectiveDate(event.getEffectiveDate());
     fpml.getEventIdentifier().add(FpMLHelper.makeBusinessEventIdentifier(contract.getPartyReference().getHref()));
     LoanContractReference ref = FpMLHelper.factory.createLoanContractReference();
